tests: fix off-by-one error in preprocessor tests
This is currently not an issue; however, if we would like to use a larger dataset for testing at some point, sklearn's train_test_split might split the data in such a way which will cause the tests to fail. For example, we have a dataset of size 3663 records, with a validation set fraction of 0.2. The training set in this case should have 3663 * (1 - 0.2) = 2930.4 records. However, train_test_split selects 2929 records, which makes the tests fail as 2930.4 and 2929 have more than a difference of 1, which is the currently set absolute tolerance. Increasing the absolute tolerance in the assertions when comparing training and validation set sizes to 2 should solve this potential problem.

Signed-off-by: Salman Maqbool salman.maqbool@cern.ch